scancode-toolkit/requirements.txt
Flask
requests
pyotp
PyJWT
